Rename NetworkStats component to runPreCallTest component.
